Add scripted build
The intel build instructions are phenomenally complicated and would have you
building an entire kernel and grub.efi.  For the purposes of the secure boot
reference platform, we don't really care what the EFI payload is, so this
scripted build simply adds the EFI Shell as the payload (meaning it won't work
with the Arduino IDE).  The build process is also coded for the Galileo Gen 1
(Kipps Bay Rev D).  If you have a Gen 2, you'll have to alter the last part
where the platform data is added.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>

Conflicts:
	.gitignore