Fix Secure Boot build
First, there's a build failure.  The inclusion of SecureBootConfig.inf isn't
allowed because embedded has no configuration menus

Then we hit asserts on running, so remove them.  The first one is a simple
ASSERT(FALSE) that Intel put into the code to prevent anyone running secure
boot.  The next is an assert on a now non-existent Quark lockdown variable and
the final few are asserting the presence of various secure boot keys we don't
want.

Finally we need to patch edk2 to compute hashes correctly on IA 32 and add
secure boot to the build and away we go.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>