Add a fake CapsuleLib implementation.
Add a capsule implementation that determines what capsules are available
by iterating over variables with the guid gFakeCapsuleHeaderGuid (
32e678e9-263f-4eae-a239-38a3cad6a1b5 ) and allows updates of any
reasonably looking data it finds.  Each valid variable contains one
EFI_SYSTEM_RESOURCE_ENTRY - i.e. the individual entry that would go in
the EFI System Resource Table (ESRT).

When UpdateCapsule() is called, each of these variables is iterated
looking for one containing a FwClass GUID matching the Capsule header
guid.  When found, version numbers are checked against the
LowestSupportedFwVersion in the previously stored entry, and the flags
of the new data are compared to those of the update.  If both are
correct, the FwVersion in the variable is updated.  Whether they're
correct or not, LastAttemptVersion and LastAttemptStatus are updated in
the saved variable, and Capsule####/CapsuleLast variables are adjusted
as appropriate.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Peter Jones <pjones@redhat.com>