Add a fake ESRT driver to go with Fake Capsule driver.
This patch adds a generator for ESRT data.  To create the ESRT, a table
is allocated during DXE and filled with the current set of updates
which would be accepted by the FakeCapsuleLib driver.

With this patch, we have a full (if minimal) framework for testing the
/other/ side of this API; that is, you can deploy a full stack of
production-level OS and pre-boot firmware update utilities with no
special code /inside/ those utilities.  To do so you have to augment
them with some system for setting variables in the
gFakeCapsuleHeaderGuid GUID-space, and something to drive your test
harness (i.e. simulated user input at some level.)

A sample (and somewhat simplified) workflow is something like this:

   [tester]
    - create {gFakeCapsuleHeaderGuid} variables
    - populate fs with updates
    - sets up OS code to be run when we boot up, so we can test it
    - boots the VM
 ,>[Ovmf] <-'  (that's a curved little arrow, you see, from the previous
 |              thing doing work to the current thing doing work)
 |  - lots of stuff  -----,
 | [Ovmf / FakeEsrtDxe] <-'
 |  - GetVariable({gFakeCapsuleHeaderGuid} variables)
 |  - InstallConfigurationTable(ESRT)
 | [Ovmf] <-'
 |  - loads OS
 | [OS] <-'
 |  - reads the status of updates available from ESRT
 |  - looks for Capsule####/CapsuleLast/CapsuleMax for previous state
 |  - reads update files it may have available
 |  - makes some decisions from the above
 |  - decides there's an update!
 |  - sets some state and a BootNext variable to start a pre-boot
 |    application to do firmware updates
 |  - reboots
 | [Ovmf] <-'
 |  - GetVariable({gFakeCapsuleHeaderGuid} variables)
 |  - InstallConfigurationTable(ESRT)
 |  - pre-boot application (fwupdate.efi)
 | [fwupdate.efi] <-'
 |  - Looks for state left from OS utility
 |  - reads the firmware files and coalesces them into a scatter gather
 |    list with capsule headers as appropriate
 |  - calls UpdateCapsule() ---,
 | [Ovmf / FakeCapsuleLib ] <--'
 |  - SetVariable() on the appropriate {gFakeCapsuleHeaderGuid} variables
 *  - updates ESRT maybe
 |\-- calls EfiResetSystem() maybe
 |  - returns to caller maybe
 | [fwupdate.efi] <--'
 |  - calls EfiResetSystem()
 +------------'

And then at some point [tester] examines according to some criteria to
decide if the thing being tested worked.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Peter Jones <pjones@redhat.com>