IntelFrameworkModulePkg: Add a EFI_CAPSULE_ARRAY_HEADER definition.
EFI_CAPSULE_TABLE is an implementation detail, and it shouldn't be
defined in UefISpec.h, but to remove it and not make "git bisect" break,
while still keeping in line with the edk2 custom of not changing
multiple packages, we have to add it with a new name here before
removing it over there.

This adds it into the implementation at the only place it's used.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Peter Jones <pjones@redhat.com>